Find daemon process when started differently
This commit will search from the term "flextimed" at the end of the
string from the proc entries.  This means that the flextime print
command will find the process even when started in different ways than
from the command line.

For example, I now start the daemon from the Gnome startup application
thing.