Fix buffer overflow issue
This commit will properly declare the name of the FIFO named pipe
file handle so that it can contain the entire file name.

This was a potential buffer overflow problem if an attacker could
control the HOME environment variable.  It also caused SELinux to
terminate the daemon.  (It did not start on Fedora 32.)