Terminate daemon process in signal handler
This commit will fix an issue where the daemon process would not
properly terminate when a terminating signal was received.  The
problem was that we where using the standard `exit` function, instead
of `_exit`.

The reason that running `exit` in a signal handler will not work is
that it is not safe as documented in signal-safety(7).

https://stackoverflow.com/a/55179208
https://man7.org/linux/man-pages/man7/signal-safety.7.html

In general, our approach to flushing data in a signal handler is
wrong.  If the main program is in the process of writing data to
files, then since the signal handler is also writing to files, the
state of the allocated data buffers are simply wrong.

We need to fix that, and avoid using stdio functions from the signal
handler.  In fact, only functions listed in signal-safety(7) can be
used in a signal handler.