Do not write to named pipe unless needed
This commit will avoid writing data to the named pipe unless it is
responding to a signal (meaning that there is a print command that is
reading the pipe.)

Again, this indicates that named pipes are perhaps not the right
choice for this situation.  Maybe we should go with sockets intead?