Add comment that the daemon code is wrongheaded
I read up on some unrelated stuff and found an article that argued
that no-one should write code that fork a background process and
terminate the original process.  Instead, we should let the invoker
decide if the process should run synchronously or asynchronosouly.

Second-guessing myself, I coded up the fork solution since I am old
and was not aware of the thinking above, and that I needed the syslog
logging part (which does not require the forking)

This commit will just add a comment that links to a summarizing Stack
Overflow answer.  It will not make any changes to the code.