Add option to set stop measurement
This commit will add a new option, --stop, that adds a stop
measurement.  This can be used to set a stop measurement for a day
that is _later_ that the existing measurement.

Note that it does not work to set a stop measurement that is _earlier_
than an existing measurment.  That is for a future commit.