Support stop measurements combined with --stop arg
This commit will fix the scenario where there are already stop
measurements in the data for a day, in combination with stop
measurments from the command line.

Starting with commit fb6e45a, we would break printing the day as soon
as we saw a stop measurement (the commit also introduced adding stop
measurements at the command line).  However, I failed to realize that
there are of course stop measurements in the actual data, not just
from the command line.

The purpose of the --stop argument is to allow the user to ask the
question "what if a stopped working at a specific time?", and the
previous implementation would only allow that if the that time was
earlier than any stop measurement in the actual data.

For testing purposes, this commit also adds two new options:

--no-local-data   Do not read data from disk (only command line)
--mark            Same as --stop, but measurement data (kind value 3)