Measure idle time using D-Bus if available
This patch adds support for D-Bus.  If available at compile time,
D-Bus is used to measure idle time instead of X11.  The reason is that
on a Wayland system, X11 is typically not available.  Since D-Bus is
available on most systems, and definately on mine, using D-Bus makes
more sense.