Remove build dependency on X11
This patch will remove the build and runtime dependency on X11.  In
particular, we now require X11 /or/ D-Bus at build time.  The README
file has been updated to clarify this fact.

This means that there is no need to install any X11 packages (the
libXScrnSaver-devel package) on Fedora 39.  So a standard Fedora 39
Wayland installation only requires the protobuf and glib2 packages at
build time.